update examples to new chat and task api
