+++ interfaces/kalosm/examples/chat-constrained.rs
@@ -16,14 +16,14 @@ async fn main() {
 
     // Create a chat session with the default chat model with the parser as constraints
     let model = Llama::new_chat().await.unwrap();
-    let mut chat = Chat::builder(model)
-        .with_constraints(move |_history| parser.clone())
-        .with_system_prompt("The assistant will act like a pirate. You will respond with either something you do or something you say. Respond with JSON in the format { \"type\": \"Say\", \"data\": \"hello\" } or { \"type\": \"Do\", \"data\": \"run away\" }")
-        .build();
+    let mut chat = model
+        .chat()
+        .with_system_prompt("The assistant will act like a pirate. You will respond with either something you do or something you say. Respond with JSON in the format { \"type\": \"Say\", \"data\": \"hello\" } or { \"type\": \"Do\", \"data\": \"run away\" }");
 
     // Then chat with the session
     loop {
-        chat.add_message(prompt_input("\n> ").unwrap())
+        chat(&prompt_input("\n> ").unwrap())
+            .with_constraints(parser.clone())
             .to_std_out()
             .await
             .unwrap();
