+++ interfaces/kalosm/examples/chat-mistral-2.rs
@@ -7,12 +7,12 @@ async fn main() {
         .build()
         .await
         .unwrap();
-    let mut chat = Chat::builder(model)
-        .with_system_prompt("The assistant will act like a pirate")
-        .build();
+    let mut chat = model
+        .chat()
+        .with_system_prompt("The assistant will act like a pirate");
 
     loop {
-        chat.add_message(prompt_input("\n> ").unwrap())
+        chat(&prompt_input("\n> ").unwrap())
             .to_std_out()
             .await
             .unwrap();
