+++ interfaces/kalosm/examples/chat-with-character.rs
@@ -10,16 +10,14 @@ async fn main() {
     let model = Llama::new_chat().await.unwrap();
     // Create constraints that parses (Responding as <character_name>) and then stops on the end of the assistant's response
     let constraints = LiteralParser::new(format!("(Responding as {}) ", character_name))
-        .then(model.default_assistant_constraints().unwrap());
+        .then(model.default_assistant_constraints());
     // Create a chat session with the model and the constraints
-    let mut chat = Chat::builder(model)
-        .with_system_prompt(character_description)
-        .with_constraints(move |_history| constraints.clone())
-        .build();
+    let mut chat = model.chat().with_system_prompt(character_description);
 
     // Chat with the user
     loop {
-        let mut output_stream = chat.add_message(prompt_input("\n> ").unwrap());
+        let mut output_stream =
+            chat(&prompt_input("\n> ").unwrap()).with_constraints(constraints.clone());
         print!("Bot: ");
         output_stream.to_std_out().await.unwrap();
     }
