+++ interfaces/kalosm/examples/generate-text.rs
@@ -4,15 +4,9 @@ use kalosm::language::*;
 async fn main() {
     tracing_subscriber::fmt::init();
     let prompt = "The following is a 300 word essay about why the capital of France is Paris:";
-    let mut stream = Llama::new()
-        .await
-        .unwrap()
-        .stream_text(prompt)
-        .with_max_length(300)
-        .await
-        .unwrap();
+    let llm = Llama::new().await.unwrap();
 
     print!("{}", prompt);
 
-    stream.to_std_out().await.unwrap();
+    llm(prompt).to_std_out().await.unwrap();
 }
