+++ interfaces/kalosm/examples/mistral.rs
@@ -1,5 +1,3 @@
-use std::io::Write;
-
 use kalosm::language::*;
 
 #[tokio::main]
@@ -12,11 +10,5 @@ async fn main() {
     let prompt = "The following is a 300 word essay about why the capital of France is Paris:";
     print!("{}", prompt);
 
-    let stream = llm.stream_text(prompt).with_max_length(300).await.unwrap();
-
-    let mut sentences = stream.words();
-    while let Some(text) = sentences.next().await {
-        print!("{}", text);
-        std::io::stdout().flush().unwrap();
-    }
+    llm(prompt).to_std_out().await.unwrap();
 }
