+++ interfaces/kalosm/examples/remote-open-ai-compatable.rs
@@ -10,10 +10,12 @@ async fn main() {
     tracing_subscriber::fmt::init();
 
     let base_url = std::env::var("OPENAI_API_BASE").expect("Custom OPENAI_API_BASE not set");
-    let llm = Gpt3_5::builder().with_base_url(&base_url).build();
-    let prompt = "The following is a 300 word essay about why the capital of France is Paris:";
+    let model = std::env::var("OPENAI_API_MODEL").expect("Custom OPENAI_API_MODEL not set");
+    let client = OpenAICompatibleClient::new().with_base_url(base_url);
+    let llm = OpenAICompatibleChatModel::builder().with_client(client).with_model(model).build();
+    let prompt = "Write a 300 word essay about why the capital of France is Paris";
     print!("{}", prompt);
 
-    let mut stream = llm.stream_text(prompt).with_max_length(300).await.unwrap();
-    stream.to_std_out().await.unwrap();
+    let mut chat = llm.chat();
+    chat(&prompt).to_std_out().await.unwrap();
 }
