+++ interfaces/kalosm/examples/self-chat.rs
@@ -6,17 +6,17 @@ use std::io::Write;
 #[tokio::main]
 async fn main() {
     let model = Llama::new_chat().await.unwrap();
-    let mut agent1 = Chat::builder(model.clone())
-        .with_system_prompt("The assistant will act like a pirate.")
-        .build();
-    let mut agent2 = Chat::builder(model)
-        .with_system_prompt("The assistant is curious and will ask questions about the world.")
-        .build();
+    let mut agent1 = model
+        .chat()
+        .with_system_prompt("The assistant will act like a pirate.");
+    let mut agent2 = model
+        .chat()
+        .with_system_prompt("The assistant is curious and will ask questions about the world.");
 
     let mut response = String::from("Is there anything you want to know about the world?");
     loop {
         println!("User:");
-        let mut stream = agent2.add_message(&response);
+        let mut stream = agent2(&response);
         let mut user_question = String::new();
         while let Some(token) = stream.next().await {
             print!("{token}");
@@ -26,7 +26,7 @@ async fn main() {
         println!();
 
         println!("Assistant:");
-        let mut stream = agent1.add_message(user_question);
+        let mut stream = agent1(&user_question);
         response.clear();
         while let Some(token) = stream.next().await {
             print!("{token}");
